Fixed various build warnings generated on CentOS

* remotes/origin/dev/fix-centos-build-warnings:
  tweaked release note
  Updated release notes file
  fixed cast to changing the type of the base variable
  fixed type comparison warnings
  fixed number typos
  removed unused variable
  fixed casting warning
  fixed bug in ins_data update function
  commented out unused parameter, eliminating warning
  Added more descriptive comment to hidden static variable
  fixed implicit int conversion warning
  fixed config_block_io.cxx build warning

+++ tests/ocv/test_descriptor_set.cxx
@@ -1,5 +1,5 @@
 /*ckwg +5
- * Copyright 2013 by Kitware, Inc. All Rights Reserved. Please refer to
+ * Copyright 2013-2014 by Kitware, Inc. All Rights Reserved. Please refer to
  * KITWARE_LICENSE.TXT for licensing information, or contact General Counsel,
  * Kitware, Inc., 28 Corporate Drive, Clifton Park, NY 12065.
  */
@@ -86,7 +86,7 @@ void test_conversions(const cv::Mat& data)
 #define CURRENT_TYPE "descriptor_set ("<< type_names[data.type()%8] \
                                        <<": "<<data.size() << ")"
   ocv::descriptor_set ds(data);
-  if (ds.size() != data.rows)
+  if (static_cast<int>(ds.size()) != data.rows)
   {
     TEST_ERROR(CURRENT_TYPE " is not the expected size");
   }
@@ -99,7 +99,7 @@ void test_conversions(const cv::Mat& data)
   data.convertTo(double_data, CV_64F);
   for( unsigned i=0; i<desc.size(); ++i)
   {
-    if (desc[i]->size() != data.cols)
+    if (static_cast<int>(desc[i]->size()) != data.cols)
     {
       TEST_ERROR(CURRENT_TYPE " some descriptors do not have the correct dimension");
       break;
